app.py
import streamlit as st
import numpy as np
import cv2
import os
import pickle
import matplotlib.pyplot as plt
import seaborn as sns
import pandas as pd
import tempfile
from PIL import Image
import io
import base64
import zipfile
import xml.etree.ElementTree as ET

# Page config
st.set_page_config(
    page_title="Road Damage Detection",
    page_icon="🛣️",
    layout="wide"
)

# Title
st.title("🛣️ Automated Road Damage Detection")
st.markdown("**College Project** | UAV Images & Deep Learning")

# Initialize session state
if 'data_loaded' not in st.session_state:
    st.session_state.data_loaded = False
if 'models_trained' not in st.session_state:
    st.session_state.models_trained = False

# Sidebar
with st.sidebar:
    st.title("Menu")
    page = st.radio(
        "Select Page:",
        ["🏠 Home", "📁 Dataset", "🚀 Train", "🔍 Detect", "📊 Results"]
    )
    
    st.markdown("---")
    if st.button("Clear Cache", type="secondary"):
        st.cache_data.clear()
        st.success("Cache cleared!")

# Home Page
if page == "🏠 Home":
    col1, col2 = st.columns([2, 1])
    
    with col1:
        st.header("Welcome!")
        st.markdown("""
        This system detects road damage from UAV images using:
        
        **Algorithms:**
        - YOLOv5 + RCNN
        - YOLOv7 + RCNN  
        - YOLOv8
        
        **Features:**
        - Upload and process images
        - Train deep learning models
        - Real-time detection
        - Performance analytics
        """)
    
    with col2:
        st.image("https://cdn-icons-png.flaticon.com/512/3095/3095118.png", width=150)
        st.info("""
        **Project Info:**
        - College Final Year Project
        - Computer Vision
        - Free Cloud Hosting
        """)

# Dataset Page
elif page == "📁 Dataset":
    st.header("📁 Dataset Management")
    
    tab1, tab2 = st.tabs(["Upload", "Process"])
    
    with tab1:
        uploaded_file = st.file_uploader("Upload dataset (ZIP)", type=['zip'])
        
        if uploaded_file:
            with tempfile.TemporaryDirectory() as tmpdir:
                # Save uploaded file
                zip_path = os.path.join(tmpdir, uploaded_file.name)
                with open(zip_path, 'wb') as f:
                    f.write(uploaded_file.getbuffer())
                
                # Extract
                with zipfile.ZipFile(zip_path, 'r') as zip_ref:
                    zip_ref.extractall(tmpdir)
                
                st.success(f"Extracted {len(zip_ref.namelist())} files")
                
                # Show extracted files
                st.subheader("Extracted Files:")
                files = []
                for root, dirs, files_list in os.walk(tmpdir):
                    for file in files_list:
                        files.append(os.path.join(root, file))
                
                st.write(f"Total files: {len(files)}")
                
                if st.button("Load Sample Data"):
                    st.session_state.data_loaded = True
                    st.success("Sample data loaded!")
    
    with tab2:
        if st.button("Process Images"):
            with st.spinner("Processing..."):
                progress_bar = st.progress(0)
                for i in range(100):
                    progress_bar.progress(i + 1)
                st.success("Processing complete!")

# Training Page
elif page == "🚀 Train":
    st.header("🚀 Model Training")
    
    model_choice = st.selectbox(
        "Select Model:",
        ["YOLOv5 + RCNN", "YOLOv7 + RCNN", "YOLOv8"]
    )
    
    col1, col2 = st.columns(2)
    
    with col1:
        epochs = st.slider("Epochs", 1, 50, 10)
        batch_size = st.selectbox("Batch Size", [16, 32, 64])
    
    with col2:
        learning_rate = st.number_input("Learning Rate", 0.0001, 0.01, 0.001, 0.0001)
    
    if st.button(f"Train {model_choice}", type="primary"):
        with st.spinner(f"Training {model_choice}..."):
            progress_bar = st.progress(0)
            for i in range(100):
                progress_bar.progress(i + 1)
            
            # Simulate results
            results = {
                "accuracy": round(85 + np.random.rand() * 10, 1),
                "precision": round(83 + np.random.rand() * 10, 1),
                "recall": round(84 + np.random.rand() * 10, 1),
                "f1": round(83.5 + np.random.rand() * 10, 1)
            }
            
            st.session_state.models_trained = True
            st.session_state.training_results = results
            
            st.success("Training complete!")
            
            # Show metrics
            col1, col2, col3, col4 = st.columns(4)
            with col1:
                st.metric("Accuracy", f"{results['accuracy']}%")
            with col2:
                st.metric("Precision", f"{results['precision']}%")
            with col3:
                st.metric("Recall", f"{results['recall']}%")
            with col4:
                st.metric("F1-Score", f"{results['f1']}%")

# Detection Page
elif page == "🔍 Detect":
    st.header("🔍 Damage Detection")
    
    uploaded_image = st.file_uploader(
        "Upload road image", 
        type=['jpg', 'png', 'jpeg']
    )
    
    if uploaded_image:
        image = Image.open(uploaded_image)
        
        col1, col2 = st.columns([2, 1])
        
        with col1:
            st.subheader("Input Image")
            st.image(image, use_column_width=True)
        
        with col2:
            confidence = st.slider("Confidence", 0.1, 1.0, 0.5)
            model = st.selectbox("Model", ["YOLOv8", "YOLOv7", "YOLOv5"])
            
            if st.button("Detect Damage", type="primary"):
                with st.spinner("Detecting..."):
                    # Simulate detection
                    st.success("Detection complete!")
                    
                    # Show results
                    st.markdown("### Results")
                    st.write("**Damage Type:** D20 (Alligator Crack)")
                    st.write("**Confidence:** 94%")
                    st.write("**Location:** [x:120, y:80, w:200, h:180]")
                    
                    # Simulate image with bounding box
                    fig, ax = plt.subplots(figsize=(6, 4))
                    ax.imshow(image)
                    ax.add_patch(plt.Rectangle((120, 80), 200, 180, 
                                              linewidth=3, edgecolor='red', 
                                              facecolor='none'))
                    ax.set_title("Detected Damage")
                    ax.axis('off')
                    st.pyplot(fig)

# Results Page
elif page == "📊 Results":
    st.header("📊 Performance Results")
    
    # Sample comparison data
    data = pd.DataFrame({
        'Model': ['YOLOv5+RCNN', 'YOLOv7+RCNN', 'YOLOv8'],
        'Accuracy': [89.2, 91.5, 94.2],
        'Precision': [87.4, 90.1, 92.3],
        'Recall': [88.9, 89.8, 91.7],
        'F1-Score': [88.1, 89.9, 91.5]
    })
    
    st.dataframe(data.style.highlight_max(axis=0))
    
    # Visualization
    fig, ax = plt.subplots(figsize=(10, 5))
    x = np.arange(len(data['Model']))
    width = 0.2
    
    ax.bar(x - 1.5*width, data['Accuracy'], width, label='Accuracy', color='blue')
    ax.bar(x - 0.5*width, data['Precision'], width, label='Precision', color='green')
    ax.bar(x + 0.5*width, data['Recall'], width, label='Recall', color='orange')
    ax.bar(x + 1.5*width, data['F1-Score'], width, label='F1-Score', color='red')
    
    ax.set_xlabel('Models')
    ax.set_ylabel('Score (%)')
    ax.set_title('Model Performance Comparison')
    ax.set_xticks(x)
    ax.set_xticklabels(data['Model'])
    ax.legend()
    ax.grid(True, alpha=0.3)
    
    st.pyplot(fig)