requirements.txt
streamlit==1.28.0
numpy==1.24.3
opencv-python-headless==4.8.1.78
matplotlib==3.7.2
seaborn==0.12.2
pandas==2.0.3
scikit-learn==1.3.0
Pillow==10.0.0
lxml==4.9.3